node AttestBatch endpoint
