+++ node/utils.go
@@ -1,155 +1,216 @@
 package node
 
 import (
-	"bytes"
 	"context"
-	"encoding/binary"
 	"errors"
 	"fmt"
 
+	"github.com/Layr-Labs/eigenda/api"
 	pb "github.com/Layr-Labs/eigenda/api/grpc/node"
 	"github.com/Layr-Labs/eigenda/common/pubip"
 	"github.com/Layr-Labs/eigenda/core"
+	"github.com/Layr-Labs/eigenda/encoding"
+	"github.com/consensys/gnark-crypto/ecc/bn254"
+	"github.com/consensys/gnark-crypto/ecc/bn254/fp"
+	"github.com/gammazero/workerpool"
 )
 
-const (
-	// Caution: the change to these prefixes needs to handle the backward compatibility,
-	// making sure the new code work with old data in DA Node store.
-	blobHeaderPrefix      = "_BLOB_HEADER_"  // The prefix of the blob header key.
-	batchHeaderPrefix     = "_BATCH_HEADER_" // The prefix of the batch header key.
-	batchExpirationPrefix = "_EXPIRATION_"   // The prefix of the batch expiration key.
-	// blobExpirationPrefix is the prefix of the blob and blob header expiration key.
-	// The blobs/blob headers expired by this prefix are those that are not associated with any batch.
-	// All blobs/blob headers in a batch are expired by the batch expiration key above.
-	blobExpirationPrefix = "_BLOBEXPIRATION_"
-	blobPrefix           = "_BLOB_" // The prefix of the blob key.
-)
-
-// EncodeBlobKey returns an encoded key as blob identification.
-func EncodeBlobKey(batchHeaderHash [32]byte, blobIndex int, quorumID core.QuorumID) ([]byte, error) {
-	buf := bytes.NewBuffer(batchHeaderHash[:])
-	err := binary.Write(buf, binary.LittleEndian, int32(blobIndex))
-	if err != nil {
-		return nil, err
+// GetBatchHeader constructs a core.BatchHeader from a proto of pb.StoreChunksRequest.
+// Note the StoreChunksRequest is validated as soon as it enters the node gRPC
+// interface, see grpc.Server.validateStoreChunkRequest.
+func GetBatchHeader(in *pb.BatchHeader) (*core.BatchHeader, error) {
+	if in == nil || len(in.GetBatchRoot()) == 0 {
+		return nil, api.NewInvalidArgError("batch header is nil or empty")
 	}
-	err = binary.Write(buf, binary.LittleEndian, quorumID)
-	if err != nil {
-		return nil, err
+	var batchRoot [32]byte
+	copy(batchRoot[:], in.GetBatchRoot())
+	batchHeader := core.BatchHeader{
+		ReferenceBlockNumber: uint(in.GetReferenceBlockNumber()),
+		BatchRoot:            batchRoot,
 	}
-	return buf.Bytes(), nil
+	return &batchHeader, nil
 }
 
-func EncodeBlobKeyByHash(blobHeaderHash [32]byte, quorumID core.QuorumID) ([]byte, error) {
-	prefix := []byte(blobHeaderPrefix)
-	buf := bytes.NewBuffer(append(prefix, blobHeaderHash[:]...))
-	err := binary.Write(buf, binary.LittleEndian, quorumID)
-	if err != nil {
-		return nil, err
+// GetBlobMessages constructs a core.BlobMessage array from blob protobufs.
+// Note the proto request is validated as soon as it enters the node gRPC
+// interface. This method assumes the blobs are valid.
+func GetBlobMessages(pbBlobs []*pb.Blob, numWorkers int) ([]*core.BlobMessage, error) {
+	blobs := make([]*core.BlobMessage, len(pbBlobs))
+	pool := workerpool.New(numWorkers)
+	resultChan := make(chan error, len(blobs))
+	for i, blob := range pbBlobs {
+		i := i
+		blob := blob
+		pool.Submit(func() {
+			blobHeader, err := GetBlobHeaderFromProto(blob.GetHeader())
+
+			if err != nil {
+				resultChan <- err
+				return
+			}
+			if len(blob.GetBundles()) != len(blob.GetHeader().GetQuorumHeaders()) {
+				resultChan <- fmt.Errorf("number of quorum headers (%d) does not match number of bundles in blob message (%d)", len(blob.GetHeader().GetQuorumHeaders()), len(blob.GetBundles()))
+				return
+			}
+
+			format := GetBundleEncodingFormat(blob)
+			bundles := make(map[core.QuorumID]core.Bundle, len(blob.GetBundles()))
+			for j, bundle := range blob.GetBundles() {
+				quorumID := blob.GetHeader().GetQuorumHeaders()[j].GetQuorumId()
+				if format == core.GnarkBundleEncodingFormat {
+					if len(bundle.GetBundle()) > 0 {
+						bundleMsg, err := new(core.Bundle).Deserialize(bundle.GetBundle())
+						if err != nil {
+							resultChan <- err
+							return
+						}
+						bundles[uint8(quorumID)] = bundleMsg
+					} else {
+						bundles[uint8(quorumID)] = make([]*encoding.Frame, 0)
+					}
+				} else if format == core.GobBundleEncodingFormat {
+					bundles[uint8(quorumID)] = make([]*encoding.Frame, len(bundle.GetChunks()))
+					for k, data := range bundle.GetChunks() {
+						chunk, err := new(encoding.Frame).Deserialize(data)
+						if err != nil {
+							resultChan <- err
+							return
+						}
+						bundles[uint8(quorumID)][k] = chunk
+					}
+				} else {
+					resultChan <- fmt.Errorf("invalid bundle encoding format: %d", format)
+					return
+				}
+			}
+
+			blobs[i] = &core.BlobMessage{
+				BlobHeader: blobHeader,
+				Bundles:    bundles,
+			}
+
+			resultChan <- nil
+		})
 	}
-	return buf.Bytes(), nil
-}
-
-func EncodeBlobKeyByHashPrefix(blobHeaderHash [32]byte) []byte {
-	prefix := []byte(blobHeaderPrefix)
-	buf := bytes.NewBuffer(append(prefix, blobHeaderHash[:]...))
-	return buf.Bytes()
-}
-
-// EncodeBlobHeaderKey returns an encoded key as blob header identification.
-func EncodeBlobHeaderKey(batchHeaderHash [32]byte, blobIndex int) ([]byte, error) {
-	prefix := []byte(blobHeaderPrefix)
-	buf := bytes.NewBuffer(append(prefix, batchHeaderHash[:]...))
-	err := binary.Write(buf, binary.LittleEndian, int32(blobIndex))
-	if err != nil {
-		return nil, err
+	pool.StopWait()
+	close(resultChan)
+	for err := range resultChan {
+		if err != nil {
+			return nil, err
+		}
 	}
-	return buf.Bytes(), nil
+	return blobs, nil
 }
 
-// Returns an encoded prefix of blob header key.
-func EncodeBlobHeaderKeyPrefix(batchHeaderHash [32]byte) []byte {
-	prefix := []byte(blobHeaderPrefix)
-	buf := bytes.NewBuffer(append(prefix, batchHeaderHash[:]...))
-	return buf.Bytes()
-}
-
-func EncodeBlobHeaderKeyByHash(blobHeaderHash [32]byte) []byte {
-	prefix := []byte(blobHeaderPrefix)
-	buf := bytes.NewBuffer(append(prefix, blobHeaderHash[:]...))
-	return buf.Bytes()
-}
-
-// EncodeBatchHeaderKey returns an encoded key as batch header identification.
-func EncodeBatchHeaderKey(batchHeaderHash [32]byte) []byte {
-	prefix := []byte(batchHeaderPrefix)
-	buf := bytes.NewBuffer(append(prefix, batchHeaderHash[:]...))
-	return buf.Bytes()
-}
-
-// Returns the encoded prefix for batch expiration key.
-func EncodeBatchExpirationKeyPrefix() []byte {
-	return []byte(batchExpirationPrefix)
-}
-
-// Returns the encoded prefix for blob expiration key.
-func EncodeBlobExpirationKeyPrefix() []byte {
-	return []byte(blobExpirationPrefix)
-}
-
-// Returns an encoded key for expration time.
-// Note: the encoded key will preserve the order of expiration time, that is,
-// expirationTime1 < expirationTime2 <=>
-// EncodeBatchExpirationKey(expirationTime1) < EncodeBatchExpirationKey(expirationTime2)
-func EncodeBatchExpirationKey(expirationTime int64) []byte {
-	prefix := []byte(batchExpirationPrefix)
-	ts := make([]byte, 8)
-	binary.BigEndian.PutUint64(ts[0:8], uint64(expirationTime))
-	buf := bytes.NewBuffer(append(prefix, ts[:]...))
-	return buf.Bytes()
-}
-
-// Returns an encoded key for expration time for blob header hashes.
-// Note: the encoded key will preserve the order of expiration time, that is,
-// expirationTime1 < expirationTime2 <=>
-// EncodeBlobExpirationKey(expirationTime1) < EncodeBlobExpirationKey(expirationTime2)
-func EncodeBlobExpirationKey(expirationTime int64) []byte {
-	prefix := []byte(blobExpirationPrefix)
-	ts := make([]byte, 8)
-	binary.BigEndian.PutUint64(ts[0:8], uint64(expirationTime))
-	buf := bytes.NewBuffer(append(prefix, ts[:]...))
-	return buf.Bytes()
-}
-
-// Returns the expiration timestamp encoded in the key.
-func DecodeBatchExpirationKey(key []byte) (int64, error) {
-	if len(key) != len(batchExpirationPrefix)+8 {
-		return 0, errors.New("the expiration key is invalid")
-	}
-	ts := int64(binary.BigEndian.Uint64(key[len(key)-8:]))
-	return ts, nil
-}
-
-// Returns the expiration timestamp encoded in the key.
-func DecodeBlobExpirationKey(key []byte) (int64, error) {
-	if len(key) != len(blobExpirationPrefix)+8 {
-		return 0, errors.New("the expiration key is invalid")
-	}
-	ts := int64(binary.BigEndian.Uint64(key[len(key)-8:]))
-	return ts, nil
-}
-
-func DecodeHashSlice(input []byte) ([][32]byte, error) {
-	if len(input)%32 != 0 {
-		return nil, errors.New("input length is not a multiple of 32")
-	}
-	numHashes := len(input) / 32
-
-	result := make([][32]byte, numHashes)
-	for i := 0; i < numHashes; i++ {
-		copy(result[i][:], input[i*32:(i+1)*32])
+func ValidatePointsFromBlobHeader(h *pb.BlobHeader) error {
+	commitX := new(fp.Element).SetBytes(h.GetCommitment().GetX())
+	commitY := new(fp.Element).SetBytes(h.GetCommitment().GetY())
+	commitment := &encoding.G1Commitment{
+		X: *commitX,
+		Y: *commitY,
 	}
 
-	return result, nil
+	if !(*bn254.G1Affine)(commitment).IsInSubGroup() {
+		return errors.New("commitment is not in the subgroup")
+	}
+
+	var lengthCommitment, lengthProof encoding.G2Commitment
+	if h.GetLengthCommitment() != nil {
+		lengthCommitment.X.A0 = *new(fp.Element).SetBytes(h.GetLengthCommitment().GetXA0())
+		lengthCommitment.X.A1 = *new(fp.Element).SetBytes(h.GetLengthCommitment().GetXA1())
+		lengthCommitment.Y.A0 = *new(fp.Element).SetBytes(h.GetLengthCommitment().GetYA0())
+		lengthCommitment.Y.A1 = *new(fp.Element).SetBytes(h.GetLengthCommitment().GetYA1())
+	}
+
+	if !(*bn254.G2Affine)(&lengthCommitment).IsInSubGroup() {
+		return errors.New("lengthCommitment is not in the subgroup")
+	}
+
+	if h.GetLengthProof() != nil {
+		lengthProof.X.A0 = *new(fp.Element).SetBytes(h.GetLengthProof().GetXA0())
+		lengthProof.X.A1 = *new(fp.Element).SetBytes(h.GetLengthProof().GetXA1())
+		lengthProof.Y.A0 = *new(fp.Element).SetBytes(h.GetLengthProof().GetYA0())
+		lengthProof.Y.A1 = *new(fp.Element).SetBytes(h.GetLengthProof().GetYA1())
+	}
+
+	if !(*bn254.G2Affine)(&lengthProof).IsInSubGroup() {
+		return errors.New("lengthProof is not in the subgroup")
+	}
+	return nil
+}
+
+// GetBlobHeaderFromProto constructs a core.BlobHeader from a proto of pb.BlobHeader.
+func GetBlobHeaderFromProto(h *pb.BlobHeader) (*core.BlobHeader, error) {
+
+	if h == nil {
+		return nil, api.NewInvalidArgError("GetBlobHeaderFromProto: blob header is nil")
+
+	}
+
+	commitX := new(fp.Element).SetBytes(h.GetCommitment().GetX())
+	commitY := new(fp.Element).SetBytes(h.GetCommitment().GetY())
+	commitment := &encoding.G1Commitment{
+		X: *commitX,
+		Y: *commitY,
+	}
+
+	if !(*bn254.G1Affine)(commitment).IsInSubGroup() {
+		return nil, errors.New("commitment is not in the subgroup")
+	}
+
+	var lengthCommitment, lengthProof encoding.G2Commitment
+	if h.GetLengthCommitment() != nil {
+		lengthCommitment.X.A0 = *new(fp.Element).SetBytes(h.GetLengthCommitment().GetXA0())
+		lengthCommitment.X.A1 = *new(fp.Element).SetBytes(h.GetLengthCommitment().GetXA1())
+		lengthCommitment.Y.A0 = *new(fp.Element).SetBytes(h.GetLengthCommitment().GetYA0())
+		lengthCommitment.Y.A1 = *new(fp.Element).SetBytes(h.GetLengthCommitment().GetYA1())
+	}
+
+	if !(*bn254.G2Affine)(&lengthCommitment).IsInSubGroup() {
+		return nil, errors.New("lengthCommitment is not in the subgroup")
+	}
+
+	if h.GetLengthProof() != nil {
+		lengthProof.X.A0 = *new(fp.Element).SetBytes(h.GetLengthProof().GetXA0())
+		lengthProof.X.A1 = *new(fp.Element).SetBytes(h.GetLengthProof().GetXA1())
+		lengthProof.Y.A0 = *new(fp.Element).SetBytes(h.GetLengthProof().GetYA0())
+		lengthProof.Y.A1 = *new(fp.Element).SetBytes(h.GetLengthProof().GetYA1())
+	}
+
+	if !(*bn254.G2Affine)(&lengthProof).IsInSubGroup() {
+		return nil, errors.New("lengthProof is not in the subgroup")
+	}
+
+	quorumHeaders := make([]*core.BlobQuorumInfo, len(h.GetQuorumHeaders()))
+	for i, header := range h.GetQuorumHeaders() {
+		if header.GetQuorumId() > core.MaxQuorumID {
+			return nil, api.NewInvalidArgError(fmt.Sprintf("quorum ID must be in range [0, %d], but found %d", core.MaxQuorumID, header.GetQuorumId()))
+		}
+		if err := core.ValidateSecurityParam(header.GetConfirmationThreshold(), header.GetAdversaryThreshold()); err != nil {
+			return nil, err
+		}
+
+		quorumHeaders[i] = &core.BlobQuorumInfo{
+			SecurityParam: core.SecurityParam{
+				QuorumID:              core.QuorumID(header.GetQuorumId()),
+				AdversaryThreshold:    uint8(header.GetAdversaryThreshold()),
+				ConfirmationThreshold: uint8(header.GetConfirmationThreshold()),
+				QuorumRate:            header.GetRatelimit(),
+			},
+			ChunkLength: uint(header.GetChunkLength()),
+		}
+	}
+
+	return &core.BlobHeader{
+		BlobCommitments: encoding.BlobCommitments{
+			Commitment:       commitment,
+			LengthCommitment: &lengthCommitment,
+			LengthProof:      &lengthProof,
+			Length:           uint(h.GetLength()),
+		},
+		QuorumInfos: quorumHeaders,
+		AccountID:   h.AccountId,
+	}, nil
 }
 
 func SocketAddress(ctx context.Context, provider pubip.Provider, dispersalPort string, retrievalPort string) (string, error) {
@@ -175,3 +236,28 @@ func GetBundleEncodingFormat(blob *pb.Blob) core.BundleEncodingFormat {
 	}
 	return core.GobBundleEncodingFormat
 }
+
+// // Constructs a core.SecurityParam from a proto of pb.SecurityParams.
+// func GetSecurityParam(p []*pb.SecurityParam) []*core.SecurityParam {
+// 	res := make([]*core.SecurityParam, len(p))
+// 	for i := range p {
+// 		res[i] = &core.SecurityParam{
+// 			QuorumID:           core.QuorumID(p[i].GetQuorumId()),
+// 			AdversaryThreshold: uint8(p[i].GetAdversaryThreshold()),
+// 		}
+// 	}
+// 	return res
+// }
+
+// // Constructs a core.QuorumParam array from a proto of pb.BatchHeader.
+// func GetQuorumParams(p *pb.BatchHeader) []core.QuorumParam {
+// 	quorum := make([]core.QuorumParam, 0)
+// 	for _, param := range p.GetQuorumParams() {
+// 		qp := core.QuorumParam{
+// 			QuorumID:        core.QuorumID(param.GetQuorumId()),
+// 			ConfirmationThreshold: uint8(param.GetQuorumThreshold()),
+// 		}
+// 		quorum = append(quorum, qp)
+// 	}
+// 	return quorum
+// }
